+++ eslint-config-module/rules/import.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // Duplicates base functionality
+  'import/no-duplicates': 'off',
+}