+++ stylelint-config/CHANGELOG.md
@@ -1,5 +1,22 @@
 # 📅 Stylelint Config Changelog
 
+## 4.3.0 - 2023-08-23
+
+* Updated dependencies
+
+## 4.2.1 - 2023-07-24
+
+* Removed stylistic to resolve missing module error
+
+## 4.2.0 - 2023-07-24
+
+* Updated dependencies
+
+## 4.1.0 - 2023-06-01
+
+* Added `overflow: scroll` to disallowed property values list
+* Updated dependencies
+
 ## 4.0.0 - 2023-05-02
 
 * Added `stylelint-stylistic` to keep soon to be removed stylistic rules